Refactor `SolverSemiImplicit` - reorg body, contact, muscle, particle kernels into separate files - standardize kernel and kernel calling fn names
