+++ newton/_src/solvers/semi_implicit/solver_semi_implicit.py
@@ -18,18 +18,21 @@ import warp as wp
 from ...core.types import override
 from ...sim import Contacts, Control, Model, State
 from ..solver import SolverBase
-from .kernels import (
-    eval_bending_forces,
-    eval_body_contact_forces,
+from .kernels_body import (
     eval_body_joint_forces,
-    eval_muscle_forces,
+)
+from .kernels_contact import (
+    eval_body_contact_forces,
     eval_particle_body_contact_forces,
-    eval_spring_forces,
-    eval_tetrahedral_forces,
+    eval_particle_contact_forces,
     eval_triangle_contact_forces,
+)
+from .kernels_particle import (
+    eval_bending_forces,
+    eval_spring_forces,
+    eval_tetrahedra_forces,
     eval_triangle_forces,
 )
-from .particles import eval_particle_forces
 
 
 class SolverSemiImplicit(SolverBase):
@@ -89,8 +92,8 @@ class SolverSemiImplicit(SolverBase):
         self,
         state_in: State,
         state_out: State,
-        control: Control,
-        contacts: Contacts,
+        control: Control | None,
+        contacts: Contacts | None,
         dt: float,
     ):
         with wp.ScopedTimer("simulate", False):
@@ -114,21 +117,25 @@ class SolverSemiImplicit(SolverBase):
             # triangle elastic and lift/drag forces
             eval_triangle_forces(model, state_in, control, particle_f)
 
-            # triangle/triangle contacts
-            if self.enable_tri_contact:
-                eval_triangle_contact_forces(model, state_in, particle_f)
-
             # triangle bending
             eval_bending_forces(model, state_in, particle_f)
 
             # tetrahedral FEM
-            eval_tetrahedral_forces(model, state_in, control, particle_f)
+            eval_tetrahedra_forces(model, state_in, control, particle_f)
 
             # body joints
             eval_body_joint_forces(model, state_in, control, body_f, self.joint_attach_ke, self.joint_attach_kd)
 
+            # muscles
+            if False:
+                eval_muscle_forces(model, state_in, control, body_f)
+
             # particle-particle interactions
-            eval_particle_forces(model, state_in, particle_f)
+            eval_particle_contact_forces(model, state_in, particle_f)
+
+            # triangle/triangle contacts
+            if self.enable_tri_contact:
+                eval_triangle_contact_forces(model, state_in, particle_f)
 
             # body contacts
             eval_body_contact_forces(model, state_in, contacts, friction_smoothing=self.friction_smoothing)
@@ -138,12 +145,8 @@ class SolverSemiImplicit(SolverBase):
                 model, state_in, contacts, particle_f, body_f, body_f_in_world_frame=False
             )
 
-            # muscles
-            if False:
-                eval_muscle_forces(model, state_in, control, body_f)
+            self.integrate_particles(model, state_in, state_out, dt)
 
             self.integrate_bodies(model, state_in, state_out, dt, self.angular_damping)
-
-            self.integrate_particles(model, state_in, state_out, dt)
 
             return state_out
